Fix LLM Spans missing Trace ID (#1)
* Remove fallback for `active_span_id` lookup

The `active_span_id` method will now rely solely on the span ID managed internally. This removes the previous fallback mechanism that would attempt to retrieve the span ID from the global Datadog tracing context, which occasionally caused spans to appear isolated in Datadog and not be linked to a trace.

* Bump version